fix(sidecar): address PR #268 review — honest GPU tier + robustness
- load(): reject the cuda plan when torch has no CUDA runtime, so FunASR's
  silent CPU fallback can't mislabel a CPU run as cuda (resolver falls back to
  the labelled cpu plan). Maps only generic "cuda" → cuda:0.
- catalog: gpu-cuda compute_type float16 → float32 (the model runs fp32; ~940MB
  VRAM). Honest label rather than unenforced fp16.
- setup.sh: opt-in CUDA torch via TORCH_PACKAGES/TORCH_INDEX_URL (CPU default
  kept) so the gpu-cuda tier is actually reachable from a provisioned env.
- transcribe(): guard empty funasr output → AsrResult("", None).
- tests: cuda-unavailable rejection, cpu-unaffected, empty-result; delenv
  SOKUJI_ASR_REPO in download-specs test (flaky under env override); accurate
  smoke skip reason.
- spec: sync status + tag-stripping (regex, not rich_transcription_postprocess)
  + float32.

Declined (with reasoning on PR): anchor tag-strip (strip-all is intentional for
guaranteed tag-free output); rename SOKUJI_ASR_REPO (reuse is intentional; a
stale value now fails loudly via the cuda/repo load guard).

+++ sidecar/sokuji_sidecar/catalog.py
@@ -37,7 +37,7 @@ ASR_MODELS: list[AsrModel] = [
                          "AEmotionStudio/cohere-transcribe-03-2026-models", 1.0),),
              recommended=True, sort_order=0),
     AsrModel("sense-voice", "SenseVoice", ("zh", "en", "ja", "ko", "yue"),
-             (Deployment("funasr_sensevoice", "gpu-cuda", "float16", SENSE_VOICE_REPO, 1.0),
+             (Deployment("funasr_sensevoice", "gpu-cuda", "float32", SENSE_VOICE_REPO, 1.0),
               Deployment("funasr_sensevoice", "cpu", "float32", SENSE_VOICE_REPO, 1.0)),
              recommended=True, sort_order=1),
     AsrModel("whisper-tiny", "Whisper tiny", ("multi",),
